Move whisper tests to own files+Add tests that fail based on party interactions
